\name{bigvideo-package}
\alias{bigvideo-package}
\docType{package}
\title{
Bigvideo.
}
\description{
Short description.
}
\details{
\tabular{ll}{
Package: \tab bigvideo\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2010-06-07\cr
License: \tab LGPL-3\cr
Copyright: \tab (C) 2010 John W. Emerson and Taylor Arnold and Michael J. Kane\cr
URL: \tab http://www.bigmemory.org\cr
LazyLoad: \tab yes\cr
}

More details.

}
\author{
John W. Emerson, Taylor Arnold, and Michael J. Kane

Maintainers: Mike and Jay <bigmemoryauthors@gmail.com>
}
\keyword{package}
\references{The Bigmemory Project: \url{http://www.bigmemory.org/}.}
\seealso{For example, see
\code{\link[bigmemory]{big.matrix}}.
}
\examples{
# None provided here.
}
