\name{probecamera}
\alias{probecamera}
\title{
Probe a camera (if attached) for resolution.
}
\description{
Longer description.
}
\usage{
probecamera()
}
\arguments{
This might ultimately need to take an indicator for color/grayscale
or other things desired?
} 
\details{
Longer details.
}
\value{
A list with class \code{videofeed}.
}
\author{John W. Emerson, Taylor Arnold, and Michael J. Kane <bigmemoryauthors@gmail.com>}
\seealso{\code{\link[bigmemory]{big.matrix}}, \code{\link{tapply}},
\code{\link{table}}, \code{\link{split}}}
\examples{

# None yet.

}
\keyword{methods}
