\name{bigvideo-package}
\alias{bigvideo-package}
\docType{package}
\title{
Bigvideo.
}
\description{
See \code{\link{bigvideo}}, \code{\link{bigtable}}, \code{\link{bigsplit}},
and \code{\link{bigtsummary}}.
}
\details{
\tabular{ll}{
Package: \tab bigvideo\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2010-06-07\cr
License: \tab LGPL-3\cr
Copyright: \tab (C) 2010 John W. Emerson and Taylor Arnold and Michael J. Kane\cr
URL: \tab http://www.bigmemory.org\cr
LazyLoad: \tab yes\cr
}

The \pkg{bigmemory} package contains the core support and basic summary
functions; \pkg{bigvideo} contains extended tabulation functionality.

}
\author{
Michael J. Kane and John W. Emerson

Maintainers: Mike and Jay <bigmemoryauthors@gmail.com>
}
\keyword{package}
\references{The Bigmemory Project: \url{http://www.bigmemory.org/}.}
\seealso{For example, see
\code{\link[bigmemory]{big.matrix}}, \code{\link[biglm]{biglm}},
\code{\link[biganalytics]{bigkmeans}}, \code{\link[biganalytics]{binit}}, \code{\link{tapply}}.
}
\examples{
# None provided here.
}
